Updating Folder Structure and source code
