+++ pyproject.toml
@@ -48,6 +48,9 @@ files = ["requirements.txt"]
 [tool.hatch.version]
 source = "vcs"
 
+[tool.hatch.build.hooks.vcs]
+version-file = "src/pokie/_version.py"
+
 [tool.hatch.version.raw-options]
 local_scheme = "no-local-version"
